Revert "Adding NUGET_XPROJ_WRITE_TARGETS to allow writing out targets and props for xproj"
This reverts commit 1adb60b665b3316c05e715cf2c777e92d94ec133.
