+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/RestoreProjectJsonTest.cs
@@ -1987,7 +1987,7 @@ namespace NuGet.CommandLine.Test
                 var packageAPath = Path.Combine("fallback2", "packagea", "1.1.0-beta-01", "build", "uap", "packageA.targets");
 
                 // B is installed to the user folder
-                var packageBPath = "$(NuGetPackageRoot)"
+                var packageBPath = "$(NuGetPackageRoot)" 
                     + Path.DirectorySeparatorChar
                     + Path.Combine("packageb", "2.2.0-beta-02", "build", "uap", "packageB.targets");
 
@@ -2295,218 +2295,6 @@ namespace NuGet.CommandLine.Test
                 // Assert
                 Assert.True(0 == r.Item1, r.Item2 + " " + r.Item3);
                 Assert.False(File.Exists(targetFilePath));
-            }
-        }
-
-        [Fact]
-        public void RestoreProjectJson_GenerateTargetsForXProjProjectsWithFlagSet()
-        {
-            // Arrange
-            using (var workingPath = TestFileSystemUtility.CreateRandomTestFolder())
-            {
-                string folderName = Path.GetFileName(workingPath);
-
-                var repositoryPath = Path.Combine(workingPath, "Repository");
-                var nugetexe = Util.GetNuGetExePath();
-
-                Directory.CreateDirectory(repositoryPath);
-                Directory.CreateDirectory(Path.Combine(workingPath, ".nuget"));
-                Util.CreateConfigForGlobalPackagesFolder(workingPath);
-                var packageA = Util.CreateTestPackageBuilder("packageA", "1.1.0-beta-01");
-                var packageB = Util.CreateTestPackageBuilder("packageB", "2.2.0-beta-02");
-
-                var targetContent = "<?xml version=\"1.0\" encoding=\"utf-8\"?><Project ToolsVersion=\"12.0\" xmlns=\"http://schemas.microsoft.com/developer/msbuild/2003\"></Project>";
-
-                var targetA = Util.CreatePackageFile("build/uap/packageA.targets", targetContent);
-                var libA = Util.CreatePackageFile("lib/uap/a.dll", "a");
-
-                packageA.Files.Add(targetA);
-                packageA.Files.Add(libA);
-
-                var targetB = Util.CreatePackageFile("build/uap/packageB.targets", targetContent);
-                var libB = Util.CreatePackageFile("lib/uap/b.dll", "b");
-
-                packageB.Files.Add(targetB);
-                packageB.Files.Add(libB);
-
-                Util.CreateTestPackage(packageA, repositoryPath);
-                Util.CreateTestPackage(packageB, repositoryPath);
-
-                Util.CreateFile(workingPath, "project.json",
-                                                @"{
-                                                    'dependencies': {
-                                                    'packageA': '1.1.0-beta-*',
-                                                    'packageB': '2.2.0-beta-*'
-                                                    },
-                                                    'frameworks': {
-                                                                'uap10.0': { }
-                                                            }
-                                                  }");
-
-                Util.CreateFile(workingPath, "test.xproj", Util.GetCSProjXML("test"));
-
-                string[] args = new string[] {
-                    "restore",
-                    "-Source",
-                    repositoryPath,
-                    "-solutionDir",
-                    workingPath,
-                    "test.xproj"
-                };
-
-                var targetFilePath = Path.Combine(workingPath, "test.nuget.targets");
-
-                // Act
-                var r = CommandRunner.Run(
-                    nugetexe,
-                    workingPath,
-                    string.Join(" ", args),
-                    waitForExit: true,
-                    environmentVariables: new Dictionary<string, string>()
-                    {
-                        {  "NUGET_XPROJ_WRITE_TARGETS", "true" }
-                    });
-
-                // Assert
-                Assert.True(0 == r.Item1, r.Item2 + " " + r.Item3);
-                Assert.True(File.Exists(targetFilePath));
-            }
-        }
-
-        [Fact]
-        public void RestoreProjectJson_MultipleFrameworksForCSProjGeneratesTargets_SameTargets()
-        {
-            // Arrange
-            using (var workingPath = TestFileSystemUtility.CreateRandomTestFolder())
-            {
-                string folderName = Path.GetFileName(workingPath);
-
-                var repositoryPath = Path.Combine(workingPath, "Repository");
-                var nugetexe = Util.GetNuGetExePath();
-
-                Directory.CreateDirectory(repositoryPath);
-                Directory.CreateDirectory(Path.Combine(workingPath, ".nuget"));
-                Util.CreateConfigForGlobalPackagesFolder(workingPath);
-                var packageA = Util.CreateTestPackageBuilder("packageA", "1.0.0");
-
-                var targetContent = "<?xml version=\"1.0\" encoding=\"utf-8\"?><Project ToolsVersion=\"12.0\" xmlns=\"http://schemas.microsoft.com/developer/msbuild/2003\"></Project>";
-
-                var targetARoot = Util.CreatePackageFile("build/packageA.targets", targetContent);
-
-                packageA.Files.Add(targetARoot);
-
-                Util.CreateTestPackage(packageA, repositoryPath);
-
-                Util.CreateFile(workingPath, "project.json",
-                                                @"{
-                                                    'dependencies': {
-                                                       'packageA': '1.0.0',
-                                                    },
-                                                    'frameworks': {
-                                                        'netstandard1.0': { },
-                                                        'netstandard1.5': { }
-                                                  }
-                                               }");
-
-                Util.CreateFile(workingPath, "test.csproj", Util.GetCSProjXML("test"));
-
-
-                string[] args = new string[] {
-                    "restore",
-                    "-Source",
-                    repositoryPath,
-                    "-solutionDir",
-                    workingPath,
-                    "test.csproj"
-                };
-
-                var targetFilePath = Path.Combine(workingPath, "test.nuget.targets");
-
-                // Act
-                var r = CommandRunner.Run(
-                    nugetexe,
-                    workingPath,
-                    string.Join(" ", args),
-                    waitForExit: true);
-
-                var messages = r.Item2 + " " + r.Item3;
-
-                // Assert
-                Assert.True(0 == r.Item1, r.Item2 + " " + r.Item3);
-                Assert.True(File.Exists(targetFilePath));
-                Assert.DoesNotContain("cannot be fully installed in projects targeting multiple frameworks", messages);
-
-                var targetsFile = File.OpenText(targetFilePath).ReadToEnd();
-                // Verify the target was added
-                Assert.True(targetsFile.IndexOf(@"build\packageA.targets") > -1);
-            }
-        }
-
-        // Verify that if different targets are selected for the same project between frameworks, that
-        // a warning is shown and no targets are added.
-        [Fact]
-        public void RestoreProjectJson_MultipleFrameworksForCSProjGeneratesTargets_DifferentTargets()
-        {
-            // Arrange
-            using (var workingPath = TestFileSystemUtility.CreateRandomTestFolder())
-            {
-                string folderName = Path.GetFileName(workingPath);
-
-                var repositoryPath = Path.Combine(workingPath, "Repository");
-                var nugetexe = Util.GetNuGetExePath();
-
-                Directory.CreateDirectory(repositoryPath);
-                Directory.CreateDirectory(Path.Combine(workingPath, ".nuget"));
-                Util.CreateConfigForGlobalPackagesFolder(workingPath);
-                var packageA = Util.CreateTestPackageBuilder("packageA", "1.0.0");
-
-                var targetContent = "<?xml version=\"1.0\" encoding=\"utf-8\"?><Project ToolsVersion=\"12.0\" xmlns=\"http://schemas.microsoft.com/developer/msbuild/2003\"></Project>";
-
-                var targetAns10 = Util.CreatePackageFile("build/netstandard1.0/packageA.targets", targetContent);
-                var targetAns13 = Util.CreatePackageFile("build/netstandard1.3/packageA.targets", targetContent);
-
-                packageA.Files.Add(targetAns10);
-                packageA.Files.Add(targetAns13);
-
-                Util.CreateTestPackage(packageA, repositoryPath);
-
-                Util.CreateFile(workingPath, "project.json",
-                                                @"{
-                                                    'dependencies': {
-                                                       'packageA': '1.0.0',
-                                                    },
-                                                    'frameworks': {
-                                                        'netstandard1.0': { },
-                                                        'netstandard1.5': { }
-                                                  }
-                                               }");
-
-                Util.CreateFile(workingPath, "test.csproj", Util.GetCSProjXML("test"));
-
-
-                string[] args = new string[] {
-                    "restore",
-                    "-Source",
-                    repositoryPath,
-                    "-solutionDir",
-                    workingPath,
-                    "test.csproj"
-                };
-
-                var targetFilePath = Path.Combine(workingPath, "test.nuget.targets");
-
-                // Act
-                var r = CommandRunner.Run(
-                    nugetexe,
-                    workingPath,
-                    string.Join(" ", args),
-                    waitForExit: true,
-                    timeOutInMilliseconds: -1);
-
-                // Assert
-                Assert.True(0 == r.Item1, r.Item2 + " " + r.Item3);
-                Assert.True(File.Exists(targetFilePath));
-                Assert.Contains("cannot be fully installed in projects targeting multiple frameworks", File.ReadAllText(targetFilePath));
             }
         }
